requirements.txt
discord.py
ffmpeg-python
python-dotenv
yt-dlp
aiohttp
pydantic
fastapi
youtube-search
uvicorn
ffmpeg-python